Split off ExcludeRandom into its own file (#654)
