+++ examples/des/blend.py
@@ -175,43 +175,3 @@ galsim.config.RegisterExtraOutput('blend_noiseonly',
                                   write_func = galsim.fits.writeMulti)
 
 
-#
-# Define the ExcludedRandom type for integer values.
-#
-
-def GenExcludedRandom(config, base, value_type):
-    """Generate a random integer within some range, but excluding some given values.
-    """
-    if 'rng' not in base:
-        raise ValueError("No base['rng'] available for type = ExcludedRandom")
-
-    rng = base['rng']
-    ud = galsim.UniformDeviate(rng)
-
-    req = { 'min' : int, 'max' : int }
-    ignore = [ 'exclude' ]  # We handle this separately.  Just tell GetAllParams not to 
-                            # raise an exception when it sees this.
-    params, safe = galsim.config.GetAllParams(config, base, req=req, ignore=ignore)
-
-    min = params['min']
-    max = params['max']
-
-    if 'exclude' in config:
-        exclude = config['exclude']
-    else:
-        exclude = []
-
-    # Also exclude max+1 just in case ud() == 1.
-    exclude.append(max+1)
-
-    # Draw values until we get one not in the exclude list.
-    while True:
-        val = int(math.floor(ud() * (max-min+1))) + min
-        if val not in exclude:
-            break
-
-    return val, False
-
-galsim.config.RegisterValueType('ExcludedRandom', GenExcludedRandom, [ int ])
-
-
